Add docker compose override file to .gitignore
As documented on https://docs.docker.com/compose/extends it's possible
to use a docker-compose.override.yml file to override/extend
configurations of a docker-compose.yml. As such, this is an excellent
way for one to perform own/local tweaks and overrides without
changing/impacting project files. This adds a line for it on .gitignore.